Restructure agent inner wrappers
